Now splits gap polygons into parts and detects when overlap checking is not needed.
Splitting the gap polygons into parts reduces the complexity of the overlap checking because
it will now only check against lines drawn in the current part as opposed to checking against
all the lines drawn in the other parts already processed.

When a gap has a smooth outline with no spikes in either the inner or outer boundaries it
can be optimised such that only the outer boundary needs to be drawn and no overlap checking
need be done. This can save a huge amount of time when the gap polygon has many points.