Turn the previous extruder off before the current extruder temperature is set.
If the previous extruder is turned off after the current extruder temperature is set and the
printer shares a single heater for all of the extruders, then the heater cools down and
the print fails.

Note that we can only make this change for when the previous extruder is completely finished
and not for just when it's going to standby as otherwise the heater timing logic is broken.