Initial implementation of variable width spiralized walls.
If a spiralized model layer is not solid, assume that the holes carve out the model to
leave walls. The width of the walls is determined and the flow rate used when printing the
spiral is modulated accordingly.

The centre line of the wall also has to be adjusted to take into account the difference in
wall width from the nominal value.