Added primer_outermost_extruder_always.
Before, it always added the prime tower on every layer for the extruder on the outside
irrespective of whether it had been the last extruder used on the previous layer (i.e.,
no extruder switch had occurred).

Now, this setting controls whether this will happen or not. When true, the behaviour is
the same as before, when false, the outermost prime tower is only output if the extruder
has just been switched.