Now processSkinPart() detects bridged skin areas and processes each area separately.
So now each region of bridge skin and non-bridge skin is passed to processTopBottom()
separately so the non-bridge regions will be printed using normal parameters and the
bridge regions will be printed using bridge parameters and each regions skin line direction
can be optimised for that region.

This appears to work OK but more refactoring is required to avoid duplicated calculations.