Put more effort into calculating the angle of bridge skin lines.
Now, it firstly finds the longest non-supported skin edge (at least 1mm long). If successful,
the skin lines will be orientated parallel to that edge.

If there are no non-supported skin edges longer than 1mm then it looks for the longest
supported edge longer than 1mm. If successful, the skin lines will be orientated at 90 deg
to that edge.

Otherwise, it concludes that the skin lines can be whatever is the default skin line direction
for the layer.