Rework Polygons.removeSmallAreas() so that it doesn't remove small holes unless requested.
Before, it removed all polygons (outlines or holes) whose area was below the threshold.

Now it only removes holes that are smaller than the threshold area if the remove_holes
parameter is true or if the small holes are contained within an outline which is itself
smaller than the threshold area.