Improve ordering of inner walls to reduce length of travels immediately before outer wall.
Before, it was possible that if some of the inner walls did not pass near to the outer wall z-seam
location, it could end up doing a long travel move immediately before the outer wall was
started even though there was an inner wall right next to the z-seam location.
Now, as long as there is a level 1 wall next to the z-seam location, it will print that
immediately before the outer wall so any underextrusion following a long travel move should
occur in the level 1 wall rather than the outer wall.