Rework skin edge support fix - now copes much better with models with sloping profiles.
When the walls on the layers above slope, the edges of the skin regions that those walls
contain do not lie on top of each other. This was producing a lot of skin edge support
walls and so put some effort into reducing the number of support walls actually printed.