Really force retraction before final travel move on layer if user wants it.
Before, it would omit the retraction if the travel distance was less than the retraction
min travel setting but this can cause blobs if the z-axis speed is very low so we make
sure that retraction will occur.

Note that the retraction will still not occur if the new layer starts at the same location
as the old layer ended as there is no final travel move in that situation.