Now use gen_ring_poly() to generate rings and fix unmatched chain ends bug.
To generate chain ends correctly it is necessary that the first chain end in each pair
is associated with a line that starts outside of the clipped region, i.e. it's passing
into the infill region rather than passing out of it. So we scan the points in the ring
poly to find the first point (if any) that is outside of the infill region and we start
our ring there.