In generateNoAirAbove() use the outer wall rather than the 2nd if only_one_wall_top is set.
When only_one_wall_top is set, the inner walls of the layer above can be inset so
that they are hidden by the next layer above. In that situation, it's better to use
the outline of the outer wall rather than the 2nd wall to determine where on the layer above
is not air.