Now constructs new Polygons that contains no short lines rather than erasing those lines.
The idea is that time will be saved because erasing a line involves shifting all the remaining
lines in the Polygons vector which could take quite a while if there are many short lines.
So creating the new Polygons containing only lines that are longer than the threshold should
be quicker.