As before, add lines unclipped if we believe that they don't hit the infill region boundary.
Should now catch cases where both ends of a line segment are within the infill area but the
line actually crosses a region outside of the infill area. It does this by calculating the
maximum length that a line segment can have without it having both ends in the infill area
but it passes through 2 sets of wall lines. Lines that are shorter than this are printed
without clipping and should never protrude through the walls and be visible. They could
still possibly intrude into a skin area that abuts the infill area but this has not been
observed yet.