In addWallLine() don't set distance_to_bridge_start to non-zero value.
This was creating unwanted coasting before a bridge that was too short to print.

However, there is still a problem here in the situation where a line contains multiple
bridge regions (that are printed as bridges). In that case, only the first bridge region
will have coasting before it, the other bridge regions won't.