Now, "exact" z-seam placement copes better with walls that don't cross the hint line.
The hint line runs from the middle of the mesh's bounding box to the z-seam hint point.
If a wall doesn't intersect with that line, use a new line that runs from the middle of
the wall's bounding box and the hint point. The z-seam created will move around depending
on how the middle of the walls' bounding boxes move with changing z. YMMV.