Shift skin edge support zig-zag infill a little to avoid bumps on layer below.
If the layer below is using a pattern such as grid or tri-hex, it will have bumps where
the lines in the same layer cross. By shifting the zig-zag infill we can avoid the bumps.

Note this does require the zig-zag infill density to be tuned to avoid the bumps.